feat: ItemVC에서 이미지를 불러올 수 없는 경우의 에러 처리 추가
