+++ Expo1900/Expo1900/View/Base.lproj/Main.storyboard
@@ -243,7 +243,7 @@
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
                             <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" placeholderIntrinsicWidth="358" placeholderIntrinsicHeight="128" translatesAutoresizingMaskIntoConstraints="NO" id="ZKS-Lm-yfc">
-                                <rect key="frame" x="16" y="60" width="358" height="128"/>
+                                <rect key="frame" x="16" y="61.333333333333329" width="358" height="125.66666666666667"/>
                             </imageView>
                             <textView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" editable="NO" textAlignment="natural" selectable="NO" translatesAutoresizingMaskIntoConstraints="NO" id="4mk-cv-xLW">
                                 <rect key="frame" x="10" y="198" width="370" height="612"/>
